heroku/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.18.0
requests==2.22.0
urllib3==1.25.7
Werkzeug==1.0.1
